requirements.txt
apache-libcloud==2.4.0
certifi==2018.11.29
chardet==3.0.4
idna==2.8
PyYAML==3.13
requests==2.21.0
urllib3==1.24.1
